Split common solr processor IT components into parent class
